+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/main/SetUpNavHeader.kt
@@ -1,0 +1,22 @@
+package jp.panta.misskeyandroidclient.ui.main
+
+import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.LifecycleOwner
+import com.google.android.material.navigation.NavigationView
+import jp.panta.misskeyandroidclient.databinding.NavHeaderMainBinding
+import jp.panta.misskeyandroidclient.ui.account.viewmodel.AccountViewModel
+
+internal class SetUpNavHeader(
+    private val navView: NavigationView,
+    private val lifecycleOwner: LifecycleOwner,
+    private val accountViewModel: AccountViewModel
+) {
+
+    operator fun invoke() {
+        DataBindingUtil.bind<NavHeaderMainBinding>(this.navView.getHeaderView(0))
+        val headerBinding =
+            DataBindingUtil.getBinding<NavHeaderMainBinding>(this.navView.getHeaderView(0))
+        headerBinding?.lifecycleOwner = lifecycleOwner
+        headerBinding?.accountViewModel = accountViewModel
+    }
+}
