+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/main/StartActivityFromNavDrawerItems.kt
@@ -1,0 +1,27 @@
+package jp.panta.misskeyandroidclient.ui.main
+
+import android.content.Intent
+import android.view.MenuItem
+import jp.panta.misskeyandroidclient.*
+import net.pantasystem.milktea.channel.ChannelActivity
+import net.pantasystem.milktea.drive.DriveActivity
+import net.pantasystem.milktea.gallery.GalleryPostsActivity
+
+internal class StartActivityFromNavDrawerItems(
+    val mainActivity: MainActivity
+) {
+    fun showNavDrawersActivityBy(item: MenuItem) {
+        val activity = when (item.itemId) {
+            R.id.nav_setting -> SettingsActivity::class.java
+            R.id.nav_drive -> DriveActivity::class.java
+            R.id.nav_favorite -> FavoriteActivity::class.java
+            R.id.nav_list -> ListListActivity::class.java
+            R.id.nav_antenna -> AntennaListActivity::class.java
+            R.id.nav_draft -> DraftNotesActivity::class.java
+            R.id.nav_gallery -> GalleryPostsActivity::class.java
+            R.id.nav_channel -> ChannelActivity::class.java
+            else -> throw IllegalStateException("未定義なNavigation Itemです")
+        }
+        mainActivity.startActivity(Intent(mainActivity, activity))
+    }
+}
