+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/main/ToolbarSetter.kt
@@ -1,0 +1,8 @@
+package jp.panta.misskeyandroidclient.ui.main
+
+import androidx.appcompat.widget.Toolbar
+
+interface ToolbarSetter {
+    fun setToolbar(toolbar: Toolbar)
+    fun setTitle(resId: Int)
+}
