+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/search/SearchTopFragment.kt
@@ -17,9 +17,9 @@ import com.wada811.databinding.dataBinding
 import dagger.hilt.android.AndroidEntryPoint
 import jp.panta.misskeyandroidclient.R
 import jp.panta.misskeyandroidclient.SearchActivity
-import jp.panta.misskeyandroidclient.ToolbarSetter
 import jp.panta.misskeyandroidclient.databinding.FragmentSearchTopBinding
 import jp.panta.misskeyandroidclient.ui.explore.ExploreFragment
+import jp.panta.misskeyandroidclient.ui.main.ToolbarSetter
 import jp.panta.misskeyandroidclient.ui.notes.view.TimelineFragment
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
